Files moved in correct location:      src - all sources (.cpp)      include - all headers (.h)      examples - all images
A single Makefile (with correct dependancies)

Hardcoded paths for headers removed from C++ sources.
