+++ examples/bayer8.png
