+++ examples/bubble8.png
