+++ include/ColorHist.h
