+++ src/ColorHist.cpp
