+++ src/Triplet.cpp
